refactor: move utility types
