+++ src/types.ts
@@ -1,0 +1,4 @@
+export type Maybe<T> = T | undefined | null;
+
+export type StringKey<T> = Extract<keyof T, string>;
+export type IdType<T> = StringKey<T>;
